refactor: package game.domain for domain objs
