+++ BotGame/src/game/domain/GameMove.java
@@ -1,6 +1,5 @@
-package game;
+package game.domain;
 
-import game.Mark;;
 
 public class GameMove {
     private Mark mark;
